+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Add `EveOnline::ESI::PublicContracts`
 * Add `EveOnline::ESI::Models::PublicContract`
+* Add `EveOnline::ESI::Models::PublicContractItem`
+* Add `EveOnline::ESI::PublicContract`
 
 ## v0.34.0